requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.4.1=pyhd8ed1ab_0
ca-certificates=2021.5.30=h033912b_0
certifi=2021.5.30=py38h50d1736_0
click=8.0.1=py38h50d1736_0
fastapi=0.68.1=pyhd8ed1ab_0
greenlet=1.1.1=py38h23ab428_0
h11=0.12.0=pyhd8ed1ab_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
ncurses=6.2=h0a44026_1
openssl=1.1.1l=h0d85af4_0
pip=21.0.1=py38hecd8cb5_0
pycodestyle=2.7.0=pyhd3eb1b0_0
pydantic=1.8.2=py38h96a0964_0
pymysql=1.0.2=pyhd8ed1ab_0
python=3.8.11=h88f2d9e_1
python-dotenv=0.19.0=pyhd8ed1ab_0
python_abi=3.8=2_cp38
readline=8.1=h9ed2024_0
setuptools=52.0.0=py38hecd8cb5_0
sqlalchemy=1.4.23=py38h96a0964_0
sqlite=3.36.0=hce871da_0
starlette=0.14.2=pyhd8ed1ab_0
tk=8.6.10=hb0a8c7a_0
typing-extensions=3.10.0.0=hd8ed1ab_0
typing_extensions=3.10.0.0=pyha770c72_0
uvicorn=0.15.0=py38h50d1736_1
wheel=0.37.0=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
